Rename to Force Saver, Remove file dirty check so save is triggered on all open files
- fix tabbing and spelling
