+++ ForceSaver.sublime-commands
@@ -1,0 +1,3 @@
+[
+  { "caption": "Force Saver: Save All", "command": "force_save_all" }
+]
